requirements.txt
-i https://pypi.org/simple
anyio==3.6.2 ; python_full_version >= '3.6.2'
click==8.1.3 ; python_version >= '3.7'
colorama==0.4.6 ; platform_system == 'Windows'
fastapi==0.95.0
greenlet==2.0.2 ; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
psutil==5.9.4
pydantic==1.10.7
pyjwt==2.6.0
sniffio==1.3.0 ; python_version >= '3.7'
sqlalchemy==2.0.9
starlette==0.26.1 ; python_version >= '3.7'
typing-extensions==4.5.0 ; python_version >= '3.7'
uvicorn==0.21.1
psycopg2==2.9.6